refactor: rename error file
