+++ src/services/item/tasks/get-public-item-task.ts
@@ -9,7 +9,7 @@ import {
 } from 'graasp';
 // local
 import { PublicItemService } from '../db-service';
-import { ItemNotFound, ItemNotPublic } from '../../../util/graasp-public-items';
+import { ItemNotFound, ItemNotPublic } from '../../../util/errors';
 import { BasePublicItemTask } from './base-public-item-task';
 
 interface ItemWithMemberships<E extends UnknownExtra> extends Item<E> {
